+++ CMake/Options.cmake
@@ -268,7 +268,7 @@ carla_option (
 carla_string_option (
   CARLA_BOOST_VERSION
   "Target boost version."
-  1.84.0
+  1.90.0
 )
 
 carla_string_option (