+++ LibCarla/source/carla/streaming/detail/tcp/Client.h
@@ -12,9 +12,9 @@
 #include "carla/streaming/detail/Token.h"
 #include "carla/streaming/detail/Types.h"
 
-#include <boost/asio/deadline_timer.hpp>
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
+#include <boost/asio/steady_timer.hpp>
 #include <boost/asio/strand.hpp>
 
 #include <atomic>
@@ -72,7 +72,7 @@ namespace tcp {
 
     boost::asio::io_context::strand _strand;
 
-    boost::asio::deadline_timer _connection_timer;
+    boost::asio::steady_timer _connection_timer;
 
     std::shared_ptr<BufferPool> _buffer_pool;