+++ LibCarla/source/carla/streaming/detail/tcp/ServerSession.cpp
@@ -36,7 +36,7 @@ namespace tcp {
       _session_id(SESSION_COUNTER++),
       _socket(io_context),
       _timeout(timeout),
-      _deadline(io_context),
+      _deadline(io_context, std::chrono::steady_clock::time_point::max()),
       _strand(io_context) {}
 
   void ServerSession::Open(
@@ -68,7 +68,7 @@ namespace tcp {
       };
 
       // Read the stream id.
-      _deadline.expires_from_now(_timeout);
+      _deadline.expires_after(_timeout.to_chrono());
       boost::asio::async_read(
           _socket,
           boost::asio::buffer(&_stream_id, sizeof(_stream_id)),
@@ -110,7 +110,7 @@ namespace tcp {
 
     log_debug("session", _session_id, ": sending message of", message->size(), "bytes");
 
-    _deadline.expires_from_now(_timeout);
+    _deadline.expires_after(_timeout.to_chrono());
     boost::asio::async_write(_socket, message->GetBufferSequence(), 
       boost::asio::bind_executor(_strand, handle_sent));
   }
@@ -120,7 +120,7 @@ namespace tcp {
   }
 
   void ServerSession::StartTimer() {
-    if (_deadline.expires_at() <= boost::asio::deadline_timer::traits_type::now()) {
+    if (_deadline.expiry() <= std::chrono::steady_clock::now()) {
       log_debug("session", _session_id, "timed out");
       Close();
     } else {