+++ LibCarla/source/test/server/test_benchmark_streaming.cpp
@@ -11,6 +11,7 @@
 #include <carla/streaming/Client.h>
 #include <carla/streaming/Server.h>
 
+#include <boost/asio/executor_work_guard.hpp>
 #include <boost/asio/post.hpp>
 
 #include <algorithm>
@@ -35,7 +36,7 @@ public:
       _client(),
       _message(make_special_message(message_size)),
       _client_callback(),
-      _work_to_do(_client_callback),
+      _work_to_do(boost::asio::make_work_guard(_client_callback)),
       _success_ratio(success_ratio) {}
 
   void AddStream() {
@@ -119,7 +120,7 @@ private:
 
   boost::asio::io_context _client_callback;
 
-  boost::asio::io_context::work _work_to_do;
+  boost::asio::executor_work_guard<boost::asio::io_context::executor_type> _work_to_do;
 
   const double _success_ratio;